update documentation and bump to version 0.2.0
