+++ build-docs.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+cargo +nightly rustdoc --all-features --open -- --cfg docsrs
