+++ src/lib.rs
@@ -15,15 +15,21 @@
 //! + `"wasm-bindgen-support"` - enable for
 //!   [wasm-bindgen](https://crates.io/crates/wasm-bindgen) on the
 //!   `wasm32-unknown-unknown` platform.
+//! + `"async"` - enable an asynchronous version of this API, and
+//!   conversion traits from synchronous to asynchronous.
+
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 mod concat_results;
 mod error;
 mod validatable;
 mod validation;
 mod validator;
+mod validator_fn;
 
 pub use concat_results::concat_results;
 pub use error::*;
 pub use validatable::*;
 pub use validation::*;
 pub use validator::*;
+pub use validator_fn::*;
