+++ src/validatable.rs
@@ -1,5 +1,9 @@
 use crate::ValidationErrors;
+
+#[cfg(feature = "async")]
 use futures::Future;
+
+#[cfg(feature = "async")]
 use std::pin::Pin;
 
 /// An item that can be validated.
@@ -19,11 +23,22 @@ pub trait Validatable<Key> {
     }
 }
 
+/// An item that can be validated asynchronously.
+///
+/// See [Validatable] for the synchronous version.
+#[cfg(feature = "async")]
+#[cfg_attr(docsrs, doc(cfg(feature = "async")))]
 pub trait AsyncValidatable<Key>
 where
     Key: 'static,
 {
+    /// Creates a future that will validate this item. The future
+    /// returns `Ok(())` if no errors were encountered, and returns
+    /// `Err(ValidationErrors)` if any errors were encountered.
     fn validate_future(&self) -> Pin<Box<dyn Future<Output = Result<(), ValidationErrors<Key>>>>>;
+    /// Creates a future that will validate this item. The future
+    /// returns an empty [ValidationErrors](ValidationErrors) if no
+    /// errors were encountered during validation.
     fn validate_future_or_empty(&self) -> Pin<Box<dyn Future<Output = ValidationErrors<Key>>>> {
         let future = self.validate_future();
         Box::pin(async move {
